life saved successfully

+++ sorts/quickSort.c
@@ -3,7 +3,7 @@
 #include <time.h>
 
 
-void swap(int *x, int* y)
+void swap(int* x, int* y)
 {
     int temp = *x;
     *x = *y;
@@ -12,19 +12,19 @@ void swap(int *x, int* y)
 
 void displayArray(int arr[], size_t size)
 {
-    for(int i = 0; i < size; ++i)
-        printf("%d\t",arr[i]);
+    for (int i = 0; i < size; ++i)
+        printf("%d\t", arr[i]);
     printf("\n");
 }
 
 unsigned int getNMinNMax(int arr[], unsigned int lb, unsigned int ub)
 {
-    unsigned int a = rand()%(ub-lb +1) + lb, b =rand()%(ub-lb +1) + lb,c = rand()%(ub-lb +1) + lb;
+    unsigned int a = rand() % (ub - lb + 1) + lb, b = rand() % (ub - lb + 1) + lb, c = rand() % (ub - lb + 1) + lb;
     // printf("%d %d %d \n", a,b,c);
     // getchar();
     // inefficient comparisons incoming, brace yourselves
-    if(arr[a] >= arr[b] && arr[a] < arr[c]) return a;
-    else if(arr[b] >= arr[a] && arr[b] < arr[c]) return b;
+    if (arr[a] >= arr[b] && arr[a] < arr[c]) return a;
+    else if (arr[b] >= arr[a] && arr[b] < arr[c]) return b;
     else return c;
 }
 
@@ -36,27 +36,28 @@ unsigned int partition(int arr[], unsigned int lb, unsigned int ub)
     // 2)neither smallest nor largest
     // swap(&arr[getNMinNMax(arr,lb,ub)], &arr[lb]);
     // 3)random
-    swap(&arr[rand()%(ub-lb +1) + lb], &arr[lb]);
+    swap(&arr[rand() % (ub - lb + 1) + lb], &arr[lb]);
     // 4)1st element(no optimisation)
-	int pivot = arr[lb];
-	unsigned int down = lb + 1, up = ub;
-	while(down <= up)
-	{
-		while(arr[down] <= pivot)
-			down++;
-		while(arr[up] > pivot)
-			up--;
-		if(down < up)
-		    swap(&arr[down], &arr[up]);
-	}
-	arr[lb] = arr[up];
-	arr[up] = pivot;
-	return up;
+    int pivot = arr[lb];
+    unsigned int down = lb + 1, up = ub;
+    while (down <= up)
+    {
+        while (arr[down] <= pivot)
+            down++;
+        while (arr[up] > pivot)
+            up--;
+        if (down < up)
+            swap(&arr[down], &arr[up]);
+    }
+    arr[lb] = arr[up];
+    arr[up] = pivot;
+    return up;
 }
 
 void quickSort(int arr[], unsigned int lb, unsigned int ub)
 {
-    while(lb < ub)
+    // life saving typecasting, ub can underflow to -1 and cause the program to break
+    while ((int)lb < (int)ub)
     {
         unsigned int pivot = partition(arr, lb, ub);
         // recurse for smaller part
@@ -75,9 +76,9 @@ void quickSort(int arr[], unsigned int lb, unsigned int ub)
 
 int main()
 {
-    int arr[] = {1,2,3,5,0,-1,-2,-3};
+    int arr[] = { 1,2,3,5,0,-1,-2,-3 };
     srand(time(NULL));
-    quickSort(arr, 0, sizeof(arr)/sizeof(int)-1);
-    displayArray(arr,sizeof(arr)/sizeof(int));
+    quickSort(arr, 0, sizeof(arr) / sizeof(int) - 1);
+    displayArray(arr, sizeof(arr) / sizeof(int));
     return 0;
 }
